Shader reflection for shaders.
Materials now serialize a shader.
Shaders are a resource that can be imported.
Shaders get reflected which makes the material editor more dynamic.
Shaders have the beginning of a hot-reloading feature that should be added to textures.
Moved around a lot of systems and made some design decisions regarding the workings of material data.

+++ Source/Engine/Resources/MaterialResources.cpp
@@ -1,8 +1,9 @@
 #include "Engine_PCH.h"
 #include "MaterialResources.h"
 
-Hail::uint64 Hail::GetMaterialSortValue(eMaterialType type, eBlendMode blend, uint32 shaderValues)
+Hail::uint64 Hail::GetMaterialSortValue(eMaterialType type, eBlendMode blend, uint64 shaderValues)
 {
 	const uint32 combinedTypeAndBlend = ((uint8)type << 0xf) | (uint8)blend; // Combine the two values to 2 bits as there are not many combinations
 	return shaderValues << 8 | combinedTypeAndBlend;
 }
+
